Reorganized the UI components into a feature-based folder structure.
