+++ Skylight/Skylight-WebUI/src/app/app.config.ts
@@ -1,4 +1,5 @@
 import { ApplicationConfig, importProvidersFrom } from '@angular/core';
+import { provideHttpClient, withFetch } from '@angular/common/http';
 import { provideRouter, withComponentInputBinding } from '@angular/router';
 import { NbEvaIconsModule } from '@nebular/eva-icons';
 import { NbSidebarModule, NbThemeModule } from '@nebular/theme';
@@ -7,6 +8,9 @@ import { routes } from './app.routes';
 
 export const appConfig: ApplicationConfig = {
   providers: [
+    provideHttpClient(
+      withFetch()
+    ),
     provideRouter(
       routes,
       withComponentInputBinding()
