+++ Skylight/Skylight-WebUI/src/pages/weather-event-page/components/index.ts
@@ -1,3 +1,1 @@
-export * from './weather-event-page.component';
-
 export * from './weather-event-page-title-card/weather-event-page-title-card.component';
