+++ Skylight/Skylight-WebUI/src/pages/weather-event-page/components/weather-event-page-title-card/weather-event-page-title-card.component.scss
