+++ Skylight/Skylight-WebUI/src/pages/weather-event-page/models/index.ts
@@ -1,0 +1,1 @@
+export * from './weather-event';
