+++ Skylight/Skylight-WebUI/src/pages/weather-event-page/models/weather-event.ts
@@ -1,0 +1,3 @@
+export class WeatherEvent {
+  
+}
