+++ Skylight/Skylight-WebUI/src/shared/components/icon-card/icon-card.component.html
