+++ Skylight/Skylight-WebUI/src/shared/components/tab-container/tab-container.component.scss
