+++ Skylight/Skylight-WebUI/src/shared/directives/darken/darken.directive.spec.ts
